Signals: -fixed time bug

+++ JobsList.h
@@ -8,6 +8,7 @@
 #include "Commands.h"
 #include <exception>
 #include <ctime>
+#include <cassert>
 
 using std::cout;
 using std::endl;
@@ -36,6 +37,9 @@ public:
         };
 
         void updateStatus(){
+            if(status == END)
+                return;
+
             int newStatus = 0;
             pid_t p = waitpid(pid, &newStatus, WNOHANG | WUNTRACED | WCONTINUED);
 
@@ -43,16 +47,20 @@ public:
             cout << "Checking status of " << pid << endl;
             cout << "newStatus: " << newStatus << endl;
             if( p != 0){
-
                 if(WIFSTOPPED(newStatus)){
                     //TODO: debug
                     cout << "update status stop" << endl;
+                    stopTime = time(nullptr);
                     status = STOP;
                 }else if(WIFEXITED(newStatus) || WTERMSIG(newStatus)){
                     cout << "update status end" << endl;
                     status = END;
+                }else if(WIFCONTINUED(newStatus)){
+                    cout << "update status continued" << endl;
+                    startTime = time(nullptr);
+                    status = RUN;
                 }else{
-                    cout << "update status run" << endl;
+                    cout << "Error: else | Function: " << __FUNCTION__ << " | Line: " << __LINE__  << endl;
                     status = RUN;
                 }
             }
@@ -87,23 +95,34 @@ public:
         }
 
         void stopCmd(){
-            stopTime = time(nullptr);
-            status = STOP;
-            kill(pid, SIGTSTP);
+            assert(status == RUN); // TODO: debug
+            if(status != RUN)
+                return;
+
+            updateStatus();
+            kill(pid, SIGSTOP);
         }
         void continueCmd(){
-            status = RUN; //TODO: times
+            assert(status == STOP); //TODO: debug
+            if(status != STOP)
+                return;
+
+            startTime = time(nullptr);
+            updateStatus();
             kill(pid, SIGCONT); //TODO:Check
+        }
+        void killCmd(){
+            assert(status != END); //TODO: debug
+            if(status == END)
+                return;
+
+            updateStatus();
+            kill(pid, SIGKILL);
         }
 
         cmdStatus getStatus() const {
             return status;
         }
-
-        void setStatus(cmdStatus status) {
-            JobEntry::status = status;
-        }
-
     };
     class notExist: public std::exception{
         int jobId;
@@ -115,12 +134,8 @@ public:
             return ("job-id " + std::to_string(jobId) + string(" does not exist")).c_str();
         }
     };
-    class emptyList: public std::exception{
-
-};
-    class inBG: public std::exception{
-
-    };
+    class emptyList: public std::exception{};
+    class inBG: public std::exception{};
 
 private:
     int counter;
@@ -136,17 +151,10 @@ public:
     void printJobsList();
     void killAllJobs();
     int getSize();
-    pid_t fgPid (){
-        if (fg)
-            return fg->getJobPid();
-        return 0;
-    }
+    pid_t fgPid ();
     void removeJobById(int jobId);
-
     bool contains(int jobId);
-
     void sendSigById(int sig, int jobId = 0);
-
     void bringFG(int jobId);
     void resumeOnBG(int jobId);
